拆分

+++ D3D12Initialize/D3D12Initialize/D3DUtil.h
@@ -129,8 +129,8 @@ public:
 struct SubmeshGeometry
 {
     UINT IndexCount = 0;
-    UINT StartIndexLocation = 0;
-    INT BaseVertexLocation = 0;
+    size_t StartIndexLocation = 0;
+    size_t BaseVertexLocation = 0;
 
     // Bounding box of the geometry defined by this submesh. 
     // This is used in later chapters of the book.
